feat: Basic meditation timer support
This checkpoint is the result of a bunch of messy experimentation. The
code will need significant cleanup, but it looks like it's able to show
a pausable timer and record the results in HealthKit. The launcher
screen has a bunch of duration buttons that don't do anything at this
point — still deciding whether durations should be set on the launch
screen or in the timer screen.