feat: Play a sound on start-up
This adds the code required to play a bell sound, but currently it just
plays the sound when the app starts up. Later, we'll need to hook it up
to code in the TimerRunningView, and also figure out when we're supposed
to allocate and de-allocate the memory for the sound.