feat: Duration buttons on timer running view
For now, I took out the duration button on the timer start view, since
it didn't work anyway; this also simplifies the interface. I now have
a pre-set number of buttons on the timer running view, and they work
to set the duration. Some day, they should work as toggles.